Refactor cluster uninstallation logic and update imports
Refactored cluster uninstall code, including updating package imports, replacing utility methods with functions from the `clusterutils` package, and improving type definitions with `types`. Code organization was also improved by moving files to the `pkg` directory for better structure.
